Include node preflight and up commands in all builds
